+++ detectors/autodetect/autodetect_test.go
@@ -91,6 +91,21 @@ func TestParseUnknown(t *testing.T) {
 	}
 }
 
+func TestDigitalOceanDetectorRegistered(t *testing.T) {
+	detector, err := Detector(IDDigitalOcean)
+	if err != nil {
+		t.Fatalf("got error: %v, expected no error", err)
+	}
+
+	c, ok := detector.(*composite)
+	if !ok {
+		t.Fatalf("got %T, expected composite detector", detector)
+	}
+	if len(c.detectors) != 1 {
+		t.Fatalf("got %d detectors, expected 1 detector", len(c.detectors))
+	}
+}
+
 func TestIBMCloudVPCDetectorRegistered(t *testing.T) {
 	detector, err := Detector(IDIBMCloudVPC)
 	if err != nil {